SSKW-32 Removed unnecesary page and gave that responsibility to the state.
