+++ src/client/components/pages/TemplateEditor/TemplateEditor.jsx
@@ -1,37 +1,38 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
+import { IoMdAdd as AddIcon } from 'react-icons/io';
 import MainViewProvider from '../../providers/MainViewProvider/MainViewProvider';
 import NavbarProvider from '../../providers/NavbarProvider/NavbarProvider';
 import TemplateForm from '../../organisms/TemplateForm/TemplateForm';
 import Button from '../../atoms/Button/Button';
 import { setIsAddingField, setTemplates } from '../../../redux/templates/templates.actions';
 
-const ExamCreator = (props) => {
-  const { history } = props;
+import { templatesManagement } from '../../../routes/paths';
+
+const TemplateEditor = (props) => {
+  const { history: { push } } = props;
   const dispatch = useDispatch();
 
-  const goToManagerView = () => history.push('/templates-management');
+  const goToManagerView = () => push(templatesManagement());
 
-  const templates = useSelector((state) => state.templates.templates);
+  const { templates } = useSelector((state) => state.templates);
 
   const addField = () => dispatch(setIsAddingField({ isAddingField: true }));
 
-  const addTemplate = (template) => dispatch(setTemplates({ templates: [...templates, template] }));
+  const addTemplate = (template) => {
+    dispatch(setTemplates({ templates: [...templates, template] }));
+    push(templatesManagement());
+  };
 
   return (
     <NavbarProvider>
       <MainViewProvider
         showBackButton
-        title='Crear examen'
+        title='Crear plantilla'
         showBottomLine
         onBackButtonClick={goToManagerView}
-        menu={(
-          <>
-            <Button className='--spaced'>Eliminar</Button>
-            <Button className='--spaced'>Editar</Button>
-          </>
-        )}
+        menu={<Button onClick={addField} icon={<AddIcon />}>Nuevo campo</Button>}
       >
         <TemplateForm
           onSubmit={addTemplate}
@@ -41,10 +42,10 @@ const ExamCreator = (props) => {
   );
 };
 
-ExamCreator.propTypes = {
+TemplateEditor.propTypes = {
   history: PropTypes.shape({
     push: PropTypes.func,
   }).isRequired,
 };
 
-export default ExamCreator;
+export default TemplateEditor;
