+++ src/client/routes/paths.js
@@ -1,6 +1,5 @@
 // Use the name of the component you're trying to reach
 
 export const templatesManagement = () => '/templates-management';
-export const templateCreator = () => '/templates-management/create';
-export const templateEditor = (id) => `/templates-management/${id}`;
+export const templateEditor = () => '/templates-management/editor';
 export const login = () => '/';
