+++ src/client/routes/routes.js
@@ -1,13 +1,11 @@
 import NotFound from '../components/pages/NotFound';
 import Login from '../components/pages/Login/Login';
 import TemplatesManagement from '../components/pages/TemplatesManagement/TemplatesManagement';
-import TemplateCreator from '../components/pages/TemplateCreator/TemplateCreator';
 import TemplateEditor from '../components/pages/TemplateEditor/TemplateEditor';
 
 import {
   login,
   templatesManagement,
-  templateCreator,
   templateEditor,
 } from './paths';
 
@@ -23,12 +21,7 @@ export default [
     exact: true,
   },
   {
-    path: templateCreator(),
-    component: TemplateCreator,
-    exact: true,
-  },
-  {
-    path: templateEditor(':id'),
+    path: templateEditor(),
     component: TemplateEditor,
     exact: true,
   },
